Added color property to brew phase

+++ BrewMobile/ContentParser.swift
@@ -43,7 +43,7 @@ func parseBrewState(brew: AnyObject) -> BrewState? {
 func parseBrewPhase(brew: AnyObject) -> BrewPhase? {
     if let json = brew as? JSONDictionary {
             if let min = json["min"] as AnyObject? as? Int {
-                if let temp = json["temp"] as AnyObject? as? Int {
+                if let temp = json["temp"] as AnyObject? as? Float {
                     if let tempReached = json["tempReached"] as AnyObject? as? Bool {
                         if let inProgress = json["inProgress"] as AnyObject? as? Bool {
                             var jobEnd = ""
